Forward-ported a PSD-specific legacy bugfix:
[fixes 3335546] Consider ImageMagick exit status 
of 1 valid for thumbnails. The version of ImageMagick 
on our staging/production environment treats an unknown 
TIFF field as an error, exiting with status 1, even 
though the thumbnail is created.  By accepting this 
exit status as valid we can ignore it and assume the 
thumbnail is fine.

This is an interim fix until ImageMagick is updated as 
later versions treat it as a warning.